AI_Transcription/requirements_optimized.txt
# Core transcription dependencies
openai-whisper>=20231117
faster-whisper>=1.0.0  # Optional but recommended for better performance
torch>=2.0.0
torchaudio>=2.0.0

# Audio processing
numpy>=1.24.0
soundfile>=0.12.0
pyaudio>=0.2.11

# Video/Audio downloading and processing
yt-dlp>=2023.12.30
ffmpeg-python>=0.2.0

# Web interface
streamlit>=1.28.0

# Analysis and NLP
transformers>=4.35.0
scikit-learn>=1.3.0

# Utilities
asyncio-mqtt>=0.13.0  # For async operations
aiofiles>=23.0.0      # For async file operations
python-dotenv>=1.0.0  # For environment variables

# AI APIs (optional but recommended)
openai>=1.0.0         # For GPT-powered summaries and analysis

# Optional dependencies for enhanced performance
# Install these for better performance:
# pip install faster-whisper  # Faster Whisper implementation
# pip install torch[cu121]    # CUDA support (if you have NVIDIA GPU)

# System dependencies (install via system package manager):
# - ffmpeg (required for audio processing)
# - portaudio19-dev (required for pyaudio on Linux)

# Installation notes:
# 1. For best performance, install CUDA version of PyTorch if you have NVIDIA GPU
# 2. Install ffmpeg system-wide: 
#    - macOS: brew install ffmpeg
#    - Ubuntu: sudo apt install ffmpeg
#    - Windows: Download from https://ffmpeg.org/
# 3. For faster-whisper, you may need to install additional dependencies:
#    pip install faster-whisper[gpu]  # For GPU support